data_analyst_env/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "data_analyst_env"
version = "0.1.0"
description = "Data Analyst Env"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "openenv-core>=0.1.0",
    "fastapi",
    "uvicorn",
    "pydantic",
    "openai"
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
packages = ["server"]